changed money type to decimal

+++ src/main/java/com/quicklybly/vacationcalendar/dto/CalculateRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
+
 public record CalculateRequest(@NotNull @Max(366) @PositiveOrZero Integer daysOfVacation,
-                               @NotNull @PositiveOrZero Long averageSalary) {
+                               @NotNull @PositiveOrZero BigDecimal averageSalary) {
 }
